--- generated_files/README.md
@@ -1,1 +1,0 @@
-## This is where files will ge stored locally
