+++ src/agr/app.py
@@ -1,6 +1,6 @@
 import os
 
-from vcf_file_generator import VcfFileGenerator
+from agr.vcf_file_generator import VcfFileGenerator
 
 
 host = os.environ.get('NEO4J_HOST', 'localhost')
@@ -11,7 +11,12 @@ alliance_db_version = os.environ.get('ALLIANCE_DATABASE_VERSION', 'test')
 
 uri = "bolt://" + host + ":" + str(port)
 
-if __name__ == '__main__':
-    generated_files_folder = "generated_files"
+def main(generated_files_folder="generated_files"):
     gvf = VcfFileGenerator(uri, generated_files_folder, alliance_db_version)
     gvf.generateFiles()
+
+
+if __name__ == '__main__':
+    main()
+    
+
